requirements.txt
Jinja2==2.8
Flask==0.10.1
Werkzeug==0.11.3
pony==0.6.1
waitress==0.8.10
Flask-Pystmark=0.11.1